MEME v2.0.7
